Remove Dice Result script and put it in UIManager instead
