--- Assets/Scripts/Board Script/DiceDisplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 59e5603dc606d244ab583bb5b5c321bd
